backend/requirements.txt
fastapi
uvicorn
opencv-python
mediapipe
numpy
pillow
tensorflow
scikit-learn
python-multipart